+++ NEWS.md
@@ -1,4 +1,4 @@
-# dials (development version)
+# dials 1.3.0
 
 ## Improvements